chore: move all test into a folder
